Refactored layers info code to dedicated Layers module
